requirements.txt
streamlit
PyPDF2
langchain
faiss-cpu
langchain-google-genai
google-generativeai
python-dotenv
